OT-481: UI for Flag/Unflag a message doesn't get updated (#49)
* fixes update of messages on contacts after flag/unflag (messages), verification (contacts)

* cleans up code

* implements missing cache handling for removeable cache objects triggert by Flutter world

+++ ios/Classes/DCWrapper/DCEventCallback/DCEventCallback.swift
@@ -76,7 +76,7 @@ public func handleDeltaChatEvent(event: CInt, data1: CUnsignedLong, data2: CUnsi
 
         default:
             log.info(logMessage)
-            EventChannelHandler.sharedInstance.handle(event, data1: data1, data2: data2)
+            EventChannelHandler.sharedInstance.handle(Int32(event), data1: data1, data2: data2)
     }
 
     return nil
